moved directory components back
vercel for some reason cant find my modules in components directory
