+++ components/protectedPage.js
